requirements.txt
tensorflow==2.20.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0
pydantic>=2.0.0
opencv-python
slowapi
numpy